Repo restructurization: Python package format, file formatting, add license
