+++ coordinate_structure_toolbox/__init__.py
@@ -1,0 +1,4 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+from .coordinate_structure_toolbox import CoordinateStructure
